# Tutorial: Making a Timer
In this tutorial, we will walk you through the process of building a timer using pure `HTML`, `JavaScript`, and `CSS`.

## Step 1. Create the Body of the Page

In this step, we create an `output` element to display a static number.
The `<output>` tag is used to represent the result of a calculation (like one performed by a script).